# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "大肥鱼管家后端",
    "version": "0.2.8",
    "description": "独立本地后端。通过 Bearer Token 认证；SSE 使用 fetch 流。CLI 运行成功表示执行器已结束，不等于独立验证业务目标。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4318"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/agenda": {
      "get": {
        "summary": "事务本完整快照；通知仅包含 pending/snoozed，独立于编码任务",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/calendar": {
      "get": {
        "summary": "展开未归档项目中的未完成日程；按重叠范围返回，最多366天",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgendaOccurrence"
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/notifications": {
      "get": {
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgendaNotification"
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/notifications/{id}/action": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[\\w.:-]{1,160}$"
            },
            "description": "重试复用原键及请求体；同键异请求返回409。重放返回原响应，随后重新查询最新状态。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "revision": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "acknowledge",
                      "snooze"
                    ]
                  },
                  "until": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "revision",
                  "action"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已持久化",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaNotification"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/preferences": {
      "get": {
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaPreferences"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[\\w.:-]{1,160}$"
            },
            "description": "重试复用原键及请求体；同键异请求返回409。重放返回原响应，随后重新查询最新状态。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "timezone": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "quietStart": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "quietEnd": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "timezone",
                  "quietStart",
                  "quietEnd"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已持久化",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaPreferences"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/plans/{id}/accept": {
      "post": {
        "summary": "仅用户确认后接受草案；原子创建关联待办",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[\\w.:-]{1,160}$"
            },
            "description": "重试复用原键及请求体；同键异请求返回409。重放返回原响应，随后重新查询最新状态。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "revision": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  }
                },
                "required": [
                  "revision"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已持久化",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaPlan"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/plans/{id}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[\\w.:-]{1,160}$"
            },
            "description": "重试复用原键及请求体；同键异请求返回409。重放返回原响应，随后重新查询最新状态。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "revision": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  }
                },
                "required": [
                  "revision"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已持久化",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaPlan"
                }
              }
            }
          },
          "400": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "参数无效、未认证或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本过期、幂等冲突或资源状态不允许",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "AgendaProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "maxLength": 8000,
            "default": ""
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "id",
          "revision",
          "createdAt",
          "updatedAt",
          "archived"
        ],
        "additionalProperties": false
      },
      "AgendaItem": {
        "type": "object",
        "properties": {
          "projectId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "kind": {
            "type": "string",
            "enum": [
              "todo",
              "event",
              "note"
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "notes": {
            "type": "string",
            "maxLength": 8000,
            "default": ""
          },
          "timezone": {
            "type": "string",
            "maxLength": 100
          },
          "dueAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "startsAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "endsAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "reminderAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "reminderMinutesBefore": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 525600
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "recurrence": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "daily",
                      "weekly"
                    ]
                  },
                  "interval": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 52,
                    "default": 1
                  },
                  "weekdays": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 7
                    },
                    "minItems": 1,
                    "maxItems": 7
                  },
                  "until": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "exceptions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    },
                    "maxItems": 366,
                    "default": []
                  }
                },
                "required": [
                  "frequency"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done",
              "cancelled"
            ]
          },
          "planId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "kind",
          "title",
          "timezone",
          "id",
          "revision",
          "createdAt",
          "updatedAt",
          "status",
          "planId"
        ],
        "additionalProperties": false
      },
      "AgendaPlan": {
        "type": "object",
        "properties": {
          "projectId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "notes": {
            "type": "string",
            "maxLength": 8000,
            "default": ""
          },
          "timezone": {
            "type": "string",
            "maxLength": 100
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "notes": {
                  "type": "string",
                  "maxLength": 8000,
                  "default": ""
                },
                "dueAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "default": null
                },
                "reminderAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "default": null
                }
              },
              "required": [
                "title"
              ],
              "additionalProperties": false
            },
            "minItems": 1,
            "maxItems": 60
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "accepted",
              "cancelled"
            ]
          },
          "itemIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "title",
          "timezone",
          "steps",
          "id",
          "revision",
          "createdAt",
          "updatedAt",
          "status",
          "itemIds"
        ],
        "additionalProperties": false
      },
      "AgendaNotification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemRevision": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "title": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "occurrenceAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "acknowledged",
              "snoozed",
              "cancelled"
            ]
          },
          "snoozedUntil": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "revision",
          "createdAt",
          "updatedAt",
          "itemId",
          "itemRevision",
          "title",
          "scheduledAt",
          "occurrenceAt",
          "status",
          "snoozedUntil"
        ],
        "additionalProperties": false
      },
      "AgendaPreferences": {
        "type": "object",
        "properties": {
          "timezone": {
            "type": "string",
            "maxLength": 100
          },
          "quietStart": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ]
          },
          "quietEnd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "timezone",
          "quietStart",
          "quietEnd"
        ],
        "additionalProperties": false
      },
      "AgendaOccurrence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "projectId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "reminderAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "itemId",
          "title",
          "projectId",
          "startsAt",
          "endsAt",
          "reminderAt"
        ],
        "additionalProperties": false
      },
      "AgendaSnapshot": {
        "type": "object",
        "required": [
          "projects",
          "items",
          "plans",
          "notifications",
          "preferences",
          "scheduler",
          "now"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaProject"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaItem"
            }
          },
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaPlan"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgendaNotification"
            }
          },
          "preferences": {
            "$ref": "#/components/schemas/AgendaPreferences"
          },
          "now": {
            "type": "string",
            "format": "date-time"
          },
          "scheduler": {
            "type": "object",
            "required": [
              "running",
              "lastTickAt",
              "lastError",
              "intervalMs",
              "delivery",
              "requiresBackendRunning"
            ],
            "properties": {
              "running": {
                "type": "boolean"
              },
              "lastTickAt": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "lastError": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "intervalMs": {
                "type": "integer"
              },
              "delivery": {
                "const": "frontend"
              },
              "requiresBackendRunning": {
                "const": true
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}